+++ composeApp/src/desktopMain/kotlin/ua/in/ios/devopstools/Installer.kt
@@ -45,7 +45,10 @@ data class InstallationState(
     var packageLink: String = "",
     var afterUnpackInstallFlag: Boolean = false,
     var afterUnpackInstallSudo: Boolean = false,
-    var afterUnpackInstallCmd: String = ""
+    var afterUnpackInstallCmd: String = "",
+    var afterInstallFlag: Boolean = false,
+    var afterInstallSudo: Boolean = false,
+    var afterInstallCmd: String = ""
 )
 
 /**
@@ -143,20 +146,6 @@ class GithubInstaller {
                 val assets = mutableListOf<String>()
                 val assetUrlMap = mutableMapOf<String, String>() // Map asset name to download URL
 
-//                if (release?.has("assets") == true && release.get("assets").isJsonArray) {
-//                    val assetsArray = release.getAsJsonArray("assets")
-//                    for (i in 0 until assetsArray.size()) {
-//                        val asset = assetsArray.get(i).asJsonObject
-//                        val assetName = asset.get("name").asString
-//                        assets.add(assetName)
-//
-//                        // Store download URL for each asset
-//                        if (asset.has("browser_download_url")) {
-//                            assetUrlMap[assetName] = asset.get("browser_download_url").asString
-//                        }
-//                    }
-//                }
-
                 if (release != null) {
                     if (state.parseReleaseNotes && release.has("body")) {
                         // Парсимо посилання з release notes
@@ -207,6 +196,13 @@ class GithubInstaller {
                 if (githubObj.has("asset")) {
                     state.selectedAsset = githubObj.get("asset")?.asString ?: ""
                     state.assetInstallType = githubObj.get("asset_type")?.asString ?: ""
+                }
+
+                if (githubObj.has("after_install_cmd")) {
+                    val afterInstallObj = githubObj.getAsJsonObject("after_install_cmd")
+                    state.afterInstallFlag = afterInstallObj.get("flag")?.asBoolean ?: false
+                    state.afterInstallSudo = afterInstallObj.get("sudo")?.asBoolean ?: false
+                    state.afterInstallCmd = afterInstallObj.get("cmd")?.asString ?: ""
                 }
             }
 
@@ -376,6 +372,17 @@ class GithubInstaller {
                 state.installationStatus = "Failed"
             }
 
+            if (state.afterInstallFlag) {
+                logger.d("PackageInstaller.installTask", "After install cmd flag is enabled, cmd: ${state.afterInstallCmd}")
+                if (state.afterInstallSudo) {
+                    val execCmd = "sudo ${state.afterInstallCmd}"
+                    executeCommandSudo(execCmd)
+                } else {
+                    val execCmd = "${state.afterInstallCmd}"
+                    executeCommandSudo(execCmd)
+                }
+            }
+
         } catch (e: Exception) {
             state.installationStatus = "Error: ${e.message}"
             logger.e("GithubInstaller.InstallTask", "Installation error", e)
@@ -436,6 +443,12 @@ class PackageInstaller {
                     state.afterUnpackInstallFlag = afterUnpackObj.get("flag")?.asBoolean ?: false
                     state.afterUnpackInstallSudo = afterUnpackObj.get("sudo")?.asBoolean ?: false
                     state.afterUnpackInstallCmd = afterUnpackObj.get("cmd")?.asString ?: ""
+                }
+                if (packageObj.has("after_install_cmd")) {
+                    val afterInstallObj = packageObj.getAsJsonObject("after_install_cmd")
+                    state.afterInstallFlag = afterInstallObj.get("flag")?.asBoolean ?: false
+                    state.afterInstallSudo = afterInstallObj.get("sudo")?.asBoolean ?: false
+                    state.afterInstallCmd = afterInstallObj.get("cmd")?.asString ?: ""
                 }
             }
 
@@ -650,6 +663,17 @@ class PackageInstaller {
                 state.currentVersion = getInstallTypeForTask(task)?.getCurrentVersion(task) ?: "Unknown"
             } else {
                 state.installationStatus = "Failed"
+            }
+
+            if (state.afterInstallFlag) {
+                logger.d("PackageInstaller.installTask", "After install cmd flag is enabled, cmd: ${state.afterInstallCmd}")
+                if (state.afterInstallSudo) {
+                    val execCmd = "sudo ${state.afterInstallCmd}"
+                    executeCommandSudo(execCmd)
+                } else {
+                    val execCmd = "${state.afterInstallCmd}"
+                    executeCommandSudo(execCmd)
+                }
             }
 
         } catch (e: Exception) {